(1) popularity field of Movie changed from double to int in all relevant places (2) spaces added in GeneralReport.java (3) new Order constructor for database added (4) increasePopularity added to RegisteredUser, and the use of it to NewOrderImpl

+++ VOD/src/BuildMovie.java
@@ -5,7 +5,7 @@ public interface BuildMovie {
     public int getYear();
     public String[] getActors();
     public boolean getIsAvailable();
-    public double getPopularity();
+    public int getPopularity();
     public double getRunningTime();
     public int getPrice();
 }
